# alx_api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ALX API
  version: 0.1.0
  description: >-
    Auth Core with local PostgreSQL integration coverage and least-privilege repositories.
    Legacy credentials are verified inside a restricted database function and upgraded to Argon2id on first successful login when enabled.
    Real production-user credentials have not been bulk-migrated; production password changes remain disabled until client cutover and delivery configuration.
servers:
  - url: http://127.0.0.1:3001
    description: Local development only
paths:
...
  /api/v1/orders/{id}:
    get:
      operationId: getOrder
      summary: Get an order with items and shipments
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        {
          '200':
            {
              description: Order returned,
              content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
            },
          '404':
            {
              description: Order not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
            },
        }
  /api/v1/orders/{id}/history:
    get:
      operationId: getOrderHistory
      summary: Get order audit history
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        {
          '200':
            {
              description: History returned,
              content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
            },
        }
  /api/v1/shipments:
    get:
      operationId: listShipments
      summary: List shipments
      security: [{ BearerAuth: [] }]
      responses:
        {
          '200':
            {
              description: Shipments returned,
              content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
            },
        }
...
  /api/v1/shipments/{id}/tracking:
    get:
      operationId: getShipmentTracking
      summary: Get shipment tracking events from the canonical order history
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        {
          '200':
            {
              description: Tracking returned,
              content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
            },
        }
...
  /api/v1/orders:
    post:
      operationId: createOrder
      summary: Create an order atomically with items, optional shipment, and history
      security: [{ BearerAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 200 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateOrderInput' } } }
      responses:
        '200':
          {
            description: Order created or idempotent replay,
            content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
          }
        '400':
          {
            description: Invalid input or missing idempotency key,
            content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
          }
  /api/v1/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      summary: Move an order status forward and append order history
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateOrderStatusInput' } } }
      responses:
        '200':
          {
            description: Status updated,
            content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
          }
        '409':
          {
            description: Status regression rejected,
            content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
          }
  /api/v1/shipments/{id}:
    patch:
      operationId: updateShipment
      summary: Update shipment operational fields
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ShipmentInput' } } }
      responses:
        '200':
          {
            description: Shipment updated,
            content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
          }
  /api/v1/products:
    post:
      operationId: createProduct
      summary: Create a catalog product
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductInput' } } }
      responses:
        '200':
          {
            description: Product created,
            content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
          }
  /api/v1/products/{id}:
    patch:
      operationId: updateProduct
      summary: Update a catalog product
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductInput' } } }
      responses:
        '200':
          {
            description: Product updated,
            content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } },
          }
...
  /api/v1/operations/couriers:
    post:
      operationId: createCourier
      summary: Create a courier
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
  /api/v1/operations/couriers/{id}:
    patch:
      operationId: updateCourier
      summary: Update a courier
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    delete:
      operationId: deleteCourier
      summary: Delete a courier
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
  /api/v1/operations/employees:
    post:
      operationId: createEmployee
      summary: Create an employee
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
  /api/v1/operations/employees/{id}:
    patch:
      operationId: updateEmployee
      summary: Update an employee
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    delete:
      operationId: deleteEmployee
      summary: Delete an employee
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateOrderInput:
      type: object
      required: [orderNumber, items]
      properties:
        orderNumber: { type: string, minLength: 1, maxLength: 128 }
        customerId: { type: string }
        status: { type: string }
        trackingNumber: { type: string }
        items: { type: array, maxItems: 1000, items: { $ref: '#/components/schemas/OrderItemInput' } }
        shipment: { $ref: '#/components/schemas/ShipmentInput' }
    OrderItemInput:
      type: object
      required: [quantity, unitPrice]
      properties:
        productId: { type: string }
        productName: { type: string }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, minimum: 0 }
        weight: { type: number, minimum: 0 }
        cbm: { type: number, minimum: 0 }
    ShipmentInput:
      type: object
      properties:
        trackingNumber: { type: string }
        shippingCompanyId: { type: string }
        courierId: { type: string }
        shipmentStatus: { type: string }
        shippingCost: { type: number, minimum: 0 }
        weight: { type: number, minimum: 0 }
        cartonCount: { type: integer, minimum: 0 }
    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, minLength: 1, maxLength: 80 }
        note: { type: string, maxLength: 500 }
    ProductInput:
      type: object
      required: [productId]
      properties:
        productId: { type: string, minLength: 1, maxLength: 128 }
        productNameAr: { type: string }
        productNameEn: { type: string }
        productUrl: { type: string, format: uri }
        unitPrice: { type: number, minimum: 0 }
        itemCategoryId: { type: string }
        weight: { type: number, minimum: 0 }
...
    SuccessEnvelope:
      type: object
      required: [success, data, requestId]
      properties:
        success: { type: boolean, const: true }
        data: { type: object }
        meta:
          type: object
          additionalProperties: true
        requestId: { type: string }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [message]
                properties:
                  field: { type: string }
                  message: { type: string }
            requestId: { type: string }